rename translation.ts to translate-captions
