+++ examples/translate-captions/translation-only-example.ts
