+++ package.json
@@ -47,7 +47,7 @@
     "example:summarization:compare": "npx ts-node --transpile-only --require tsconfig-paths/register examples/summarization/provider-comparison.ts",
     "example:moderation": "npx ts-node --transpile-only --require tsconfig-paths/register examples/moderation/basic-example.ts",
     "example:moderation:compare": "npx ts-node --transpile-only --require tsconfig-paths/register examples/moderation/provider-comparison.ts",
-    "example:translation": "npx ts-node --transpile-only --require tsconfig-paths/register examples/translation/basic-example.ts",
+    "example:translate-captions": "npx ts-node --transpile-only --require tsconfig-paths/register examples/translate-captions/basic-example.ts",
     "example:audio-translation": "npx ts-node --transpile-only --require tsconfig-paths/register examples/audio-translation/basic-example.ts"
   },
   "keywords": [
