+++ src/functions/index.ts
@@ -1,5 +1,5 @@
 export * from './summarization';
-export * from './translation';
+export * from './translate-captions';
 export * from './audio-translation';
 export * from './chapters';
 export * from './burned-in-captions';
